feat : upgrade software to latest versions
