+++ aws-s3-project/docker/docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - "5432:5432"
 
   localstack:
-    image: localstack/localstack:3.8.1
+    image: localstack/localstack:4.0.3
     ports:
       - "127.0.0.1:4566:4566"            # LocalStack Gateway
       - "127.0.0.1:4510-4559:4510-4559"  # external services port range
