+++ aws-s3-project/src/test/java/com/learning/awspring/common/AbstractIntegrationTest.java
@@ -4,7 +4,6 @@ import static com.learning.awspring.utils.AppConstants.PROFILE_TEST;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.learning.awspring.TestS3Application;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -12,7 +11,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
 @ActiveProfiles({PROFILE_TEST})
-@SpringBootTest(webEnvironment = RANDOM_PORT, classes = TestS3Application.class)
+@SpringBootTest(webEnvironment = RANDOM_PORT, classes = ContainerConfig.class)
 @AutoConfigureMockMvc
 public abstract class AbstractIntegrationTest {
